refactor(board): decompose board.go into granular files
- Split board.go into grid.go, plot.go, environment.go, geometry.go, and navigation.go.
- Ensure clear separation between Entities, Value Objects, and Domain Services.
- Clean up re-exports and naming consistency.
- Consolidate all tests into board_test.go.
