+++ internal/domain/board/plane.go
@@ -4,26 +4,27 @@ import (
 	"fmt"
 )
 
-// DiscoveryState représente l'état de découverte d'une zone sur la minimap
+// DiscoveryState représente l'état de découverte d'une zone sur la minimap.
 type DiscoveryState int
 
 const (
 	StateHidden   DiscoveryState = iota // Zone inconnue
-	StateAdjacent                       // Zone limitrophe non visitée
+	StateAdjacent                       // Zone limitrophe identifiée
 	StateVisited                        // Zone déjà explorée
 )
 
-// DreamPlane représente le "mega-board" composé de plusieurs zones (Grids)
+// DreamPlane représente le "mega-plateau" composé de plusieurs zones (Grids) interconnectées.
 type DreamPlane struct {
 	ID              string
 	Zones           map[string]*Grid
-	Coords          map[string]Position               // GridID -> Coords in the plane layout
+	Coords          map[string]Position             // GridID -> Coordonnées relatives dans le plan
 	Connections     map[string]map[Direction]string // GridID -> Direction -> TargetGridID
-	DiscoveryStates map[string]DiscoveryState         // GridID -> État de découverte
+	DiscoveryStates map[string]DiscoveryState       // GridID -> État de découverte
 	StartZoneID     string
 	EndZoneID       string
 }
 
+// NewDreamPlane initialise un nouveau plan de rêve vide.
 func NewDreamPlane(id string) *DreamPlane {
 	return &DreamPlane{
 		ID:              id,
@@ -34,17 +35,18 @@ func NewDreamPlane(id string) *DreamPlane {
 	}
 }
 
+// AddZone ajoute une grille au plan.
 func (p *DreamPlane) AddZone(grid *Grid) {
 	p.Zones[grid.ID] = grid
 }
 
+// Connect lie deux zones de manière bidirectionnelle.
 func (p *DreamPlane) Connect(fromID, toID string, dir Direction) {
 	if p.Connections[fromID] == nil {
 		p.Connections[fromID] = make(map[Direction]string)
 	}
 	p.Connections[fromID][dir] = toID
 
-	// Connection réciproque
 	opposite := p.OppositeDirection(dir)
 	if p.Connections[toID] == nil {
 		p.Connections[toID] = make(map[Direction]string)
@@ -52,6 +54,7 @@ func (p *DreamPlane) Connect(fromID, toID string, dir Direction) {
 	p.Connections[toID][opposite] = fromID
 }
 
+// OppositeDirection retourne l'inverse d'une direction cardinale.
 func (p *DreamPlane) OppositeDirection(dir Direction) Direction {
 	switch dir {
 	case North:
@@ -66,6 +69,7 @@ func (p *DreamPlane) OppositeDirection(dir Direction) Direction {
 	return dir
 }
 
+// GetConnectedZone retourne l'identifiant de la zone connectée dans une direction donnée.
 func (p *DreamPlane) GetConnectedZone(fromID string, dir Direction) (string, bool) {
 	if conn, ok := p.Connections[fromID]; ok {
 		targetID, ok := conn[dir]
